Fix most composed input bugs and TEMAC-1776 Three-pronged approach: 1 - I believe there was a mistake in this refactoring:  https://github.com/ianstormtaylor/slate/commit/f812816b7dcb2d4b2efa0d4ba12d4feac31850c9 Comment out the code in before plugin that "kills" beforeinput handling. This lets the after plugin work with beforeinput or input as appropriate
2 - Commented out code in before plugin that sets state on the editor
on compositionstart/end events.
This is a special case for placeholder which we do not use.
Setting state during composition seems like a bad idea in that it could break
the composition.

3 - If an input event is ever received and goes into the after plugin's
onInput(), this code does not
"know about" Leaf elements rendered from Decorations.
This means the contenteditable <span> that is edited can be much smaller
than what the code finds as the "Slate node equivalent to the edited DOM"
which then results in whacky editing.
Added a hack that does the right thing in most cases.
This mostly applies to our code syntax Decorations.

# Conflicts:
#	packages/slate-react/src/plugins/before.js