Restore error handler
If an event in Slate throws an error we want control over what happens. This gives us the opportunity to fail gracefully or react how we see fit.
This commit adds an onEventError prop to the Editor component. The function passed to that prop will be called with the error as its argument.