Enable Copying void nodes + example
This is to make it possible to copy a void inline node to the pasteboard. Inside the Emoji example there is an example of how to leverage that possibility to actually copy/paste a void inline node.
The changes in content.js make it so onCopy is fired when a selection is inside a void node like the ones in the emoji example.
Prior to Slate 47 this was enough to make copying void nodes work. However, that is not the case anymore. Updates in clone-fragment around line 63 change the way the contents of nodes are copied. Slate was using its own findDOMNode function. That was deprecated and replaced. You can find more information here:
https://github.com/ianstormtaylor/slate/commit/21d0f4a233012857d22dd95a8238bee89f25093a
https://github.com/ianstormtaylor/slate/pull/3018
The new findDOMNode works if the selection is not expanded all the way over the entire void node. I tried programmatically selecting the entire node in onCopy and verified that the selection looks the same as when it is manually selected. However, in clone-fragment they are using the native selection from window.getSelection… so while the Slate selection is correct it has not yet rendered on the screen. Because it hasn’t rendered on the screen the native selection won’t reflect the changes and therefore doesn’t populate in slate’s clone-fragment function.
This led me to look at other ways to get the inline void information onto the pasteboard.
We can create new pasteboard types in most browsers so we could add a unique pasteboard type. However, slate manually creates the transfer object that we get in onPaste. They do this to cover up browser inconsistencies and to rehydrate their custom pasteboard types. They don’t forward non-standard non-slate pasteboard types to onPaste. However, for browsers that do not support custom pasteboard types they put the unique paste options on the pasteboard as plain text and prefix them.
My primary goal with solving copy/paste for void nodes is to do so with as little change to slate as possible. If we are changing Slate code it needs to be in the form of a pull request to Slate otherwise with each change to Slate we risk big regression bugs.
Considering the above, I opted to have onCopy put plain text on the pasteboard. The plain text is prefixed. Slate uses ‘SLATE-DATA-EMBED::’ so follow in that pattern the example uses ‘EMOJI-DATA-EMBED::’. After the prefix is a string of the json that would be needed to insert the emoji as an inline node. Then, onPaste looks for the prefix. If it is found it rehydrates the json and inserts an inline node with it.